ci: Do not install and run kubernetes on Fedora
We have seen random failures when we are trying to enable kubernetes repository
in Fedora 31. There are some times that the Fedora CI is failing when we
try to import the gpgkey from the kubernetes repository
(https://packages.cloud.google.com/yum/doc/yum-key.gpg). This PR removes
that we install and run kubernetes on Fedora.

Fixes #2302

Signed-off-by: Gabriela Cervantes <gabriela.cervantes.tellez@intel.com>